MDL-35238 Remove the hide/show icon from the Plugins overview page
This was reported in forums as an usability issue. This icon is used
across Moodle as an active link to hide or show things - not as a status
icon. People were trying to disable plugins by clicking the icon. Until
there is a full support for doing this from the Plugins overview page,
it is better to remove the icons completely.