MDL-38094 question forms: remove spare repeat element blanks.
Once the question has been created, only so as many blanks as are
required by the current question. Adding a new answer / choice / unit /
etc. is a rare operation, so it is OK if it takes another mouse click.
It is better to keep the form shorter in the common case.