MDL-38756 JavaScript: Handle missing jsrev during installation
During installation, jsrev is unset so unavailable for the cache checks.
The installation sets the initial jsrev value and this is updated
periodically, unless it is set to -1 for JS developer.

During installation, we shouldn't really use any caches so if jsrev is
unset, then we should act as though it is set to developer.