MDL-41098 Atto Text editor: Add new text editor to core
This is an alternative to the TinyMCE text editor that is based on HTML5
content editable regions.

It has been designed to "gel" with Moodle and uses YUI for javascript,
Moodle standard dialogs and standard moodle buttons. It has also been designed
to work well on mobile devices and can collapse down to a narrow column if required.

Imported from the plugins database here:
https://moodle.org/plugins/view.php?plugin=editor_contenteditable

Original git repository here:
https://github.com/damyon/contenteditable